CORE-8323 refactor code.
