+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/views/dialogs/CreateFolderDialog.java
@@ -4,10 +4,13 @@ import org.iplantc.de.client.models.diskResources.Folder;
 import org.iplantc.de.client.util.DiskResourceUtil;
 import org.iplantc.de.commons.client.validators.DiskResourceNameValidator;
 import org.iplantc.de.commons.client.views.dialogs.IPlantPromptDialog;
+import org.iplantc.de.diskResource.client.DiskResourceView;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.safehtml.shared.SafeHtml;
 import com.google.gwt.user.client.ui.HTML;
+
+import com.sencha.gxt.widget.core.client.event.SelectEvent;
 
 /**
  * @author jstroot
@@ -28,19 +31,26 @@ public class CreateFolderDialog extends IPlantPromptDialog {
     private final Appearance appearance;
 
     private final DiskResourceUtil diskResourceUtil;
-    public CreateFolderDialog(final Folder parentFolder) {
+
+    public CreateFolderDialog(final Folder parentFolder, final DiskResourceView.Presenter parentPresenter) {
         this(parentFolder,
-             GWT.<Appearance> create(Appearance.class));
+             GWT.<Appearance> create(Appearance.class), parentPresenter);
     }
 
     public CreateFolderDialog(final Folder parentFolder,
-                              final Appearance appearance) {
+                              final Appearance appearance, final DiskResourceView.Presenter parentPresenter) {
         super(appearance.folderName(), -1, "", new DiskResourceNameValidator());
         this.appearance = appearance;
         diskResourceUtil = DiskResourceUtil.getInstance();
         setWidth(appearance.dialogWidth());
         setHeadingText(appearance.newFolder());
         initDestPathLabel(parentFolder.getPath());
+        addOkButtonSelectHandler(new SelectEvent.SelectHandler() {
+            @Override
+            public void onSelect(SelectEvent event) {
+                parentPresenter.doCreateNewFolder(parentFolder, getFieldText());
+            }
+        });
     }
 
     private void initDestPathLabel(String destPath) {
